Added route for fetching blogs RSS data, deleted unused code
