+++ controllers/login.js
@@ -90,56 +90,11 @@ router.all('/', (req, res, next) => {
 				$sort: 'name',
 			},
 		);
-		return schoolsPromise.then(schools => res.render('authentication/home', {
+		return schoolsPromise.then((schools) => res.render('authentication/home', {
 			schools,
-			blogFeed: [],
 			inline: true,
 			systems: [],
 		}));
-		/*
-		return new Promise((resolve) => {
-			feedr.readFeed('https://blog.schul-cloud.org/rss', {
-				requestOptions: { timeout: 2000 },
-			}, (err, data) => {
-				let blogFeed;
-				try {
-					blogFeed = data.rss.channel[0].item
-						.filter(item => (item['media:content'] || []).length && (item.link || []).length)
-						.slice(0, 3)
-						.map((e) => {
-							const date = new Date(e.pubDate);
-							const locale = 'en-us';
-							const month = date.toLocaleString(locale, { month: 'long' });
-							e.pubDate = `${date.getDate()}. ${month}`;
-							e.description = e.description.join(' ');
-							e.url = e.link[0];
-							e.img = {
-								src: e['media:content'][0].$.url,
-								alt: e.title,
-							};
-							return e;
-						});
-				} catch (e) {
-					// just catching the blog-error
-					blogFeed = [];
-				}
-				const schoolsPromise = getSelectOptions(
-					req, 'schools',
-					{
-						purpose: { $ne: 'expert' },
-						$limit: false,
-						$sort: 'name',
-					},
-				);
-				resolve(schoolsPromise.then(schools => res.render('authentication/home', {
-					schools,
-					blogFeed,
-					inline: true,
-					systems: [],
-				})));
-			});
-		});
-		*/
 	});
 });
